Plugin is now working
