+++ src/plugins/themingPlugin/actions.js
