+++ src/plugins/themingPlugin/components/IndexPage.js
