+++ tests/integration/web/hello-world/src/app/plugin.js
@@ -2,8 +2,12 @@
  * Created by umair on 8/22/17.
  */
 import InternationalizationPlugin from '../../../../../../src/plugins/InternationalizationPlugin';
+import ThemingPlugin from '../../../../../../src/plugins/themingPlugin/';
+import ReduxDevtoolsPlugin from '../../../../../../src/plugins/ReduxDevtoolsPlugin';
 
 const plugins = [
-	InternationalizationPlugin
+	InternationalizationPlugin,
+  ThemingPlugin,
+  ReduxDevtoolsPlugin
 ];
 export default plugins;
