+++ tests/integration/web/todo/src/bluerain/plugins.js
@@ -1,11 +1,10 @@
 import ReduxDevtoolsPlugin from '../../../../../../src/plugins/ReduxDevtoolsPlugin';
 import ApolloPlugin from '../../../../../../src/plugins/ApolloPlugin';
-import ThemingPlugin from '../../../../../../src/plugins/Theming/theme';
+
 
 const plugins = [
 	ReduxDevtoolsPlugin,
-	ApolloPlugin,
-	ThemingPlugin
+	ApolloPlugin
 ];
 
 export default plugins;
